Skip if docker container already exists

+++ scripts/cn/start.sh
@@ -4,8 +4,25 @@ start_phala_node()
 {
 	log_info "---------启动 phala node----------"
 	local node_name=$(cat $basedir/config.json | jq -r '.nodename')
-	docker run -ti --rm --name phala-node -d -e NODE_NAME=$node_name -p 9933:9933 -p 9944:9944 -p 30333:30333 -v $HOME/phala-node-data:/root/data swr.cn-east-3.myhuaweicloud.com/phala/phala-poc3-node
+	local ipaddr=$(cat $basedir/config.json | jq -r '.ipaddr')
+	
+	if [ ! -z $(docker ps -qf "name=phala-node") ]; then
+		log_info "---------phala node 已启动，等待同步----------"
+		while true ; do
+			local block_json=$(curl -sH "Content-Type: application/json" -d '{"id":1, "jsonrpc":"2.0", "method": "system_syncState", "params":[]}' http://$ipaddr:9933)
+			local node_block=$(echo $block_json | jq -r '.result.currentBlock')
+			local hightest_block=$(echo $block_json | jq -r '.result.highestBlock')
+			if [ x"$node_block" == x"$hightest_block" ] && [ x"$hightest_block" > x"10" ]; then
+				log_success "phala-node 同步完成"
+				break
+			fi
+			log_info "同步进度： 节点高度（$node_block），网络高度（$hightest_block）"
+			sleep 30
+		done
+		exit 0
+	fi
 
+	docker run -ti --rm --name phala-node -d -e NODE_NAME=$node_name -p 9933:9933 -p 9944:9944 -p 30333:30333 -v $HOME/phala-node-data:/root/data swr.cn-east-3.myhuaweicloud.com/phala/phala-poc3-node
 	if [ $? -ne 0 ]; then
 		log_err "----------启动 phala node 失败-------------"
 		exit 1
@@ -14,7 +31,6 @@ start_phala_node()
 	log_info "等待node节点同步区块高度"
 	sleep 30
 	while true ; do
-		local ipaddr=$(cat $basedir/config.json | jq -r '.ipaddr')
 		local block_json=$(curl -sH "Content-Type: application/json" -d '{"id":1, "jsonrpc":"2.0", "method": "system_syncState", "params":[]}' http://$ipaddr:9933)
 		local node_block=$(echo $block_json | jq -r '.result.currentBlock')
 		local hightest_block=$(echo $block_json | jq -r '.result.highestBlock')
@@ -31,32 +47,52 @@ start_phala_node_debug()
 {
 	log_info "---------启动 phala node----------"
 	local node_name=$(cat $basedir/config.json | jq -r '.nodename')
-	docker run -ti --rm --name phala-node -e NODE_NAME=$node_name -p 9933:9933 -p 9944:9944 -p 30333:30333 -v $HOME/phala-node-data:/root/data swr.cn-east-3.myhuaweicloud.com/phala/phala-poc3-node
+	local ipaddr=$(cat $basedir/config.json | jq -r '.ipaddr')
+	
+	if [ ! -z $(docker ps -qf "name=phala-node") ]; then
+		log_info "---------phala node 已启动，等待同步----------"
+		while true ; do
+			local block_json=$(curl -sH "Content-Type: application/json" -d '{"id":1, "jsonrpc":"2.0", "method": "system_syncState", "params":[]}' http://$ipaddr:9933)
+			local node_block=$(echo $block_json | jq -r '.result.currentBlock')
+			local hightest_block=$(echo $block_json | jq -r '.result.highestBlock')
+			if [ x"$node_block" == x"$hightest_block" ] && [ x"$hightest_block" > x"10" ]; then
+				log_success "phala-node 同步完成"
+				break
+			fi
+			log_info "同步进度： 节点高度（$node_block），网络高度（$hightest_block）"
+			sleep 30
+		done
+		exit 0
+	fi
 
+	docker run -ti --rm --name phala-node -e NODE_NAME=$node_name -p 9933:9933 -p 9944:9944 -p 30333:30333 -v $HOME/phala-node-data:/root/data swr.cn-east-3.myhuaweicloud.com/phala/phala-poc3-node
 	if [ $? -ne 0 ]; then
-			log_err "----------启动 phala node 失败-------------"
-			exit 1
+		log_err "----------启动 phala node 失败-------------"
+		exit 1
 	fi
 
 	log_info "等待node节点同步区块高度"
 	sleep 30
 	while true ; do
-			local ipaddr=$(cat $basedir/config.json | jq -r '.ipaddr')
-			local block_json=$(curl -sH "Content-Type: application/json" -d '{"id":1, "jsonrpc":"2.0", "method": "system_syncState", "params":[]}' http://$ipaddr:9933)
-			local node_block=$(echo $block_json | jq -r '.result.currentBlock')
-			local hightest_block=$(echo $block_json | jq -r '.result.highestBlock')
-			if [ x"$node_block" == x"$hightest_block" ] && [ x"$hightest_block" > x"10" ]; then
-					log_success "phala-node 同步完成"
-					break
-			fi
-			log_info "同步进度： 节点高度（$node_block），网络高度（$hightest_block）"
-			sleep 30
+		local block_json=$(curl -sH "Content-Type: application/json" -d '{"id":1, "jsonrpc":"2.0", "method": "system_syncState", "params":[]}' http://$ipaddr:9933)
+		local node_block=$(echo $block_json | jq -r '.result.currentBlock')
+		local hightest_block=$(echo $block_json | jq -r '.result.highestBlock')
+		if [ x"$node_block" == x"$hightest_block" ] && [ x"$hightest_block" > x"10" ]; then
+			log_success "phala-node 同步完成"
+			break
+		fi
+		log_info "同步进度： 节点高度（$node_block），网络高度（$hightest_block）"
+		sleep 30
 	done
 }
 
 start_phala_pruntime()
 {
 	log_info "----------启动pruntime----------"
+	if [ ! -z $(docker ps -qf "name=phala-pruntime") ]; then
+		log_info "---------phala-pruntime 已启动----------"
+		exit 0
+	fi
 	
 	local res_sgx=$(ls /dev | grep -w sgx)
 	local res_isgx=$(ls /dev | grep -w isgx)
@@ -78,6 +114,10 @@ start_phala_pruntime()
 start_phala_pruntime_debug()
 {
 	log_info "----------启动pruntime----------"
+	if [ ! -z $(docker ps -qf "name=phala-pruntime") ]; then
+		log_info "---------phala-pruntime 已启动----------"
+		exit 0
+	fi
 
 	local res_sgx=$(ls /dev | grep -w sgx)
 	local res_isgx=$(ls /dev | grep -w isgx)
@@ -91,18 +131,22 @@ start_phala_pruntime_debug()
 	fi
 
 	if [ $? -ne 0 ]; then
-			log_err "----------启动pruntime失败----------"
-			exit 1
+		log_err "----------启动pruntime失败----------"
+		exit 1
 	fi
 }
 
 start_phala_phost()
 {
 	log_info "----------启动phost----------"
+	if [ ! -z $(docker ps -qf "name=phala-phost") ]; then
+		log_info "---------phala-phost 已启动----------"
+		exit 0
+	fi
+
 	local ipaddr=$(cat $basedir/config.json | jq -r '.ipaddr')
 	local mnemonic=$(cat $basedir/config.json | jq -r '.mnemonic')
 	docker run -d -ti --rm --name phala-phost -e PRUNTIME_ENDPOINT="http://$ipaddr:8000" -e PHALA_NODE_WS_ENDPOINT="ws://$ipaddr:9944" -e MNEMONIC="$mnemonic" -e EXTRA_OPTS="-r" swr.cn-east-3.myhuaweicloud.com/phala/phala-poc3-phost
-
 	if [ $? -ne 0 ]; then
 		log_err "----------启动phost失败----------"
 		exit 1
@@ -111,15 +155,19 @@ start_phala_phost()
 
 start_phala_phost_debug()
 {
-        log_info "----------启动phost----------"
-        local ipaddr=$(cat $basedir/config.json | jq -r '.ipaddr')
-        local mnemonic=$(cat $basedir/config.json | jq -r '.mnemonic')
-	docker run -ti --rm --name phala-phost -e PRUNTIME_ENDPOINT="http://$ipaddr:8000" -e PHALA_NODE_WS_ENDPOINT="ws://$ipaddr:9944" -e MNEMONIC="$mnemonic" -e EXTRA_OPTS="-r" swr.cn-east-3.myhuaweicloud.com/phala/phala-poc3-phost
+	log_info "----------启动phost----------"
+	if [ ! -z $(docker ps -qf "name=phala-phost") ]; then
+		log_info "---------phala-phost 已启动----------"
+		exit 0
+	fi
 
-        if [ $? -ne 0 ]; then
-                log_err "----------启动phost失败----------"
-                exit 1
-        fi
+	local ipaddr=$(cat $basedir/config.json | jq -r '.ipaddr')
+	local mnemonic=$(cat $basedir/config.json | jq -r '.mnemonic')
+	docker run -ti --rm --name phala-phost -e PRUNTIME_ENDPOINT="http://$ipaddr:8000" -e PHALA_NODE_WS_ENDPOINT="ws://$ipaddr:9944" -e MNEMONIC="$mnemonic" -e EXTRA_OPTS="-r" swr.cn-east-3.myhuaweicloud.com/phala/phala-poc3-phost
+	if [ $? -ne 0 ]; then
+		log_err "----------启动phost失败----------"
+		exit 1
+	fi
 }
 
 start()
